requirements.txt
langchain-openai==0.1.8
langchain-core==0.2.4
langchain-community==0.2.4
chromadb==0.4.24
python-dotenv==1.0.1